Improve tests
traffic-service.spec.js: mocks the real service to prevent the depletion of the limited API calls. It checks that the service is called and checks the behaviour of the cache so it's still a useful test

db.helper.js: makes it so that the connection to the database happens only once but only returns when the server is effectively connected. Before the connection would happen multiple times because of race conditions