Modify how the traffic info is stored on the database
Now it saves every event as his own item in the db and the cache is
saved with a reference to all the relevant events.
Mongo really doesn't make joins easy however and saving the data was
annoying too. The idea was to use hooks on save and find to not have to
call .populate() every time but I'm not good enough to make that work,
I guess.

I also wanted to be able to create a new document from a populated
Object without having to deal with parsing the data in the services.
The way I found is to add a static function to the TrafficCache module
called .fromObject(). I don't think modifying the constructor is a good
idea since it's probably called after queries and I couldn't make it
work with hooks: somehow the events array was empty in pre validate and
so I couldn't save the events and get their id's; the rest of the data
was there so I'm not sure where the issue was but I'm guessing that
because the type of events in the document is different from what I want
 to pass as a parameter it was getting dumped before even the pre
 validation hook.